+++ src/chunk.c
@@ -122,6 +122,9 @@ int disassembleInstruction(Chunk* chunk, int offset) {
         case OP_BUILD_ARRAY:   return byteInstruction("OP_BUILD_ARRAY",        chunk, offset);
         case OP_GET_INDEX:     return simpleInstruction("OP_GET_INDEX",                offset);
         case OP_SET_INDEX:     return simpleInstruction("OP_SET_INDEX",                offset);
+        case OP_TRY:           return jumpInstruction("OP_TRY",          1, chunk, offset);
+        case OP_TRY_END:       return simpleInstruction("OP_TRY_END",                  offset);
+        case OP_THROW:         return simpleInstruction("OP_THROW",                    offset);
         case OP_CALL:          return byteInstruction("OP_CALL",              chunk, offset);
         case OP_RETURN:        return simpleInstruction("OP_RETURN",                offset);
         case OP_HALT:          return simpleInstruction("OP_HALT",                  offset);